Chapter3/gae-bigquery/requirements.txt
google-api-python-client==1.6.5
google-cloud-bigquery==0.31